When loading a text script, a new -n option to append a null byte has been added.
This is because when executing a text script the underlying mp function takes a null terminated string and the first thing it does is use strlen() to determine the size and then calls the lex-parser. Goes pear shaped without the null and so this new enhancement allows the null to be added.

At some point the utility will be enhanced so that -n is automatically added if the file has a .py extension.